Add scripts to compute GOS in bulk
Add a scripts folder, containing three scripts that allow to compute
GOS in bulk. Two out of three scripts depend on the Panta Rhei package
developed by CEOS for the use of an internal list of edges.

Since Panta Rhei is not yet publicly available and the same
information is in principle available from hyperspy,
the scripts should be modified in the future to use that as a
replacement.